Upgrades to fix a dependabot issue.
https://github.com/aws/eks-anywhere-packages/security/dependabot/1

Is caused by our version of github.com/docker/distribution, which is a
dependency of oras.land/oras-go. Upgrading oras required some changes to
the code, as well as our version of the helm.sh/helm/v3. That in turn
requires upgrades to a bunch of packages, including
sigs.k8s.io/controller-runtime and github.com/go-logr/logr. Those
changes required more tweaks to the code.