New command: sign-object
    go run ./cmd/sign-object --help

It has a different take on how to generate the signature from what we
currently have.

The current solution uses github.com/itchyny/gojq (a Go-library that
implements jq). I've found that operations like we're performing are
very difficult with jq. The queries are brittle and hard to read (even
harder to modify).

I started looking to see if we could use Kustomize to patch in a
signature. However, Kustomize isn't flexible enough (this is a long
story).

So for this new bit of functionality, I've gone to the library that
underlies Kustomize, jsonpatch (github.com/evanphx/json-patch). I feel
like this is more resilient, though it's definitely more lines of
code. I think this will be more maintainable and composable in the long
run. Those familiar with Kustomize should have an easier time wrapping
their heads around how this code works. Also, if, in the future, the
limitations that prevent us from using Kustomize directly are removed,
we might be able to simply do that, which would simplify a chunk of
this.

I've not functionally modified the existing signature package, though I
did perform a few tweaks that were either sanitary (extracting
constants, etc) or just made the existing code a little easier to
read/use (moving from [32]byte to []byte).

If this approach is deemed worthwhile, I would like to expand to using
its approach in the signature package. I'd also be writing a bunch of
tests of course.