Create DownloadBundle method (#141)
* Create DownloadBundle method

Extract DownloadBundle out of the LatestBundle method so that the DownloadBundle method can be re-used for bundle with different tags. Separate unit test cases have been added and all passed.

* Clean up LatestBundle method

Co-authored-by: Terry Howe <terrylhowe@gmail.com>

* Delete TestLatestBundle test

The main functionality of LatestBundle has been extracted out to a separate method DownloadBundle, which has been tested separately in TestDownloadBundle. Therefore, I propose to delete the TestLatestBundle method since it's outdated. I added a TODO to address a new validation in the future.

Co-authored-by: Terry Howe <terrylhowe@gmail.com>