Add a re-download functionality to Controller (#148)
* Add a re-download functionality to Controller

Update the Reconcile function to support the scenario that if the bundle controller detects that the active bundle is deleted, the bundle controller will validate the active bundle by namespace and name, and re-download the bundle. In order to support this functionality, I also added a GetActiveBundleNamespacedName function in the manager, and updated manager fake functions. The changes were tested locally and passed all testing.

* Refine the re-download functionality to Controller

Update Reconcile function to support the re-download logic

Co-authored-by: Terry Howe <terrylhowe@gmail.com>

* Fix syntax error

* Update error message format

Co-authored-by: Eric Wollesen <169516+ewollesen@users.noreply.github.com>

Co-authored-by: Terry Howe <terrylhowe@gmail.com>
Co-authored-by: Eric Wollesen <169516+ewollesen@users.noreply.github.com>