Sort bundles numerically (#181)
Rename the SortBundlesNewestFirst method to SortBundlesDescending. Though the current sorting logic does represent "NewestFirst", this sorting logic needs to be further tuned (and may be subject to changes). Before we re-tune the logic, we want to rename the method to a more generic name "SortBundlesDescending" to avoid confusion for users.