Extract functions and use a loop to be DRYer.
This script was doing the exact same things for two different k8s
versions. I extracted that behavior into two orthogonal functions,
then used a loop to remove the duplication.

Future improvements would be to modify the generatebundle binary so
that its output is more predictable, and make the versions in the
input files match the values used here.