Add KubeVersion check to PackageBundle State (#203)
* Update Matches method name in bundle

As we plan to create multiple matches methods in packagebundle.go, I'm renaming the current Matches method to PackageMatches to clarify the scope of the method.

* Add KubeVersionMatches method to bundle

A helper function GetMajorMinorFromString is added together to help parse the Kubernetes version name.

* Add Unit Test for KubeVersionMatches

Unit test was also added for helper function GetMajorMinorFromString. All tests passed locally.

* Fix typo

* Resolve conflicts

* Update methods / functions visibility

* Update manager to consider K8 versions

If the K8 version of the bundle doesn't match with the cluster's, the manager will mark the bundle state as PackageBundleStateIgnored.

* Fix a format issue