Extract awsAuth function and use it in more places (#211)
This extends the work performed in
https://github.com/aws/eks-anywhere-packages/pull/210 to use the
orasAuth function in more places. The calling convention has changed a
little, as the new awsAuth function no longer wraps the oras call, but
rather it calls whatever command is passed to it. This eliminates the
need for the awsAuth function to know anything about the paths where
binaries can be found.