Don't assume GetBundleList returns a sorted slice
Implements sort.Interface on slices of BundlePackage, to make it easy for
consumers that want sorted packages to be sure they have them sorted in a way
that doesn't rely on assumptions about the behavior of BundlePackage.

Modifies tests to not assume GetBundleList returns a sorted slice, but moves those
into a new test where it can be used by anyone that needs it.