Lift configfile.ConfigFile out of CredentialsStore
Rather than assuming that users of CredentialStore will know to call Init
after instantiation, the CredentialStore's dependency on a Docker *ConfigFile
is made explicit by requiring it to be passed to the constructor. Now we need
never worry that a user will forget to call Init, and it removes the awkward
pattern of New-then-Init.