Lift remote.NewRegistry out of OCIRegistryClient
Rather than assuming that users of OCIRegistryClient will know to call Init
after instantiation, the OCIRegistryClient's dependency on a
ORAS *remote.Registry is made explicit by requiring it to be passed to the
constructor. Now we need never worry that a user will forget to call Init, and
it removes the awkward pattern of New-then-Init.