Revert "Revert (#761)"
This reverts commit 360d2998571ed9bf08029afcb2c8dd7b198f5113.

The reversion of de9c0be (by 360d299) was performed due to what was believed
to be a regression, but was later found to be a config issue.

Since 360d299 squashed both the revert, plus another small change into one
commit, this revert will undo both. That is, it re-instates de9c0be, and
reverts the small change squashed into the original revert. (Clear as mud,
right?)

A follow up commit will manually re-instate the small change.