Allow eksa-controller-manager to remove package bundle controllers
These permissions will allow the eksa-controller to delete package bundle
controller helm installations completely, and without error.

The namespace in question is the workload clusters,
e.g. eksa-packages-my-workload-cluster.

The secret in question is the registry-mirror-secret that is created for
workload clusters (regardless of whether or registry mirror is in use).

Part of https://github.com/aws/eks-anywhere-packages/issues/807.