Create packages namespaces via helm
Previously, the helm chart created a specific namespace per cluster:

Self-managed clusters: "eksa-packages"
Workload clusters:     "eksa-packages-$CLUSTER_NAME"

The PBC controller was also checking for, and creating if necessary,
"eksa-packages-$CLUSTER_NAME", which would only trigger for self-managed
clusters, as the helm chart was creating it for workload clusters.

With this change, both "eksa-packages" and "eksa-packages-$CLUSTER_NAME" are
created via the helm chart. The code to create them from the PBC controller is
removed. This should simplify the entire process.